CIB-1270: Project conversion actions.  These required some major changes to action management to allow full functionality:
- actions can now have a prepare method to prepare an argument when invoked via the web UI
- actions can be enabled for template instances
- actions are not shown for invalid instances by default
- an action invocation can indicate it invalidates a path so it can be refreshed in UIs

The last item is required as conversions change the display name for the project type in the UI tree.  The implementation is just enough to not be a hack and allow this use case - it may need revisting in the future.

Also fixed bugs - incomplete custom and versioned configuration types, and state fields should not be displayed for templates.