Fixed bug restoring into HSQL.  The bug manifested as build failures due to FK violations on the recipe result node table - despite a reasonable order of inserts/updates with valid keys.  The bug persisted until Pulse was restarted twice.
Tracing into HSQL code, I found that the FK violation occured as the HSQL connection was not seeing a previously-inserted row.  Somehow it appears the state got out of whack.  Examining the db state in between restarts revealed a possible reason for the *second* restart clearing things up:  in both the initial run and the second run of Pulse, ALTER TABLE ... ADD CONSTRAINT calls were being issued.  The first time they happen in the restore itself, and the second time they are replayed from db.log (thanks to an unclean HSQL shutdown).

I am still uncertain if it is an HSQL bug (it appears so) that causes FK violations after ALTER TABLEs in the same run, but have been able to avoid this case by a full HSQL shutdown and compact post restore.  In fact this is probably a good idea in the sense that it nicely checkpoints the database after a possibly-large number of changes.  There is still a nagging concern that if this is not an HSQL bug then we were doing something illegal with our pooling of connections or similar that I don't understand...