CIB-1468: 57588: Support resetting of VMs when a build is done; CIB-1454: 57567: Should be able to set pulse agent to disable-when-idle; and CIB-1290: Allow a configurable number of ping retries before marking an agent offline.
Created a new AgentStatusManager to be responsible for maintaining the transient state of agents.  This moves some bits out of the recipe queue and the existing agent manager.  Using a new manager should make this part more testable, and also allows the control required for uniform state transitions.

Added new states to the remote agents to deal with ping races and with the different phases of a build.  Most importantly, there are new states at the end of the build where the recipe has completed but collection, cleanup and post-stage actions need to be run.  To allow actions to reset agents, they wait until after collection and cleanup and these new states indicate that the agents are still busy.  After actions the agents move into an 'awaiting ping' state so that they are not used again until a successful ping occurs.

Disable on idle is now the default if you disable an agent while it is building.  If you then ask to disable again the disable will be immediate (the build will terminate with an error).  This is supported by a new persistent enable state DISABLING.  Persistence means that an agent that is marked for disable on idle will also disable on restart.

Ping retries are supported by not immediately giving up on a failed or unexpected ping.  Instead, that ping is ignored unless we have reached some configurable timeout.  This mechanism both smooths over occasional failures and is used to handle ping races.  An example race is when we have dispatched a recipe to an agent and afterwards receive a ping saying it is idle.  This can validly happen as the ping may have happened before the recipe was started.  We distinguish between valid races and problems by using the timeout: e.g. in the previous case we expect to see a building ping (or a recipe complete ping) within the timeout.

Next up will be unit tests for the new manager and updates to the agent UI.  Although the UI can handle the new states already, it does not accurately reflect the fact that an intial disable while building will be interpreted as disable-on-idle.