Build log updates.  These fix bugs:   - logs did not have user-friendly messages, nor were the messages consistent with the style of the recipe log   - pre-build hooks were not logged   - when multiple builds ran, all build loggers captured output for all of them.
The changes also continue the move away from event handling for logging even further.  The build logger, like the recipe logger, is now driven explicitly from the build controller and both build and recipe loggers are driven by the hook manager using a new HookLogger interface.  The output events are no longer pumped with a thread, this is unnecessary with everything happening in sync and on the master.  Instead I extraced a common OutputLogger interface which is adapter fairly directly to an output stream.  Hopefully this makes things even clearer to follow.