CIB-1779: Option to show project groups collapsed by default.  Actually a complete rework of how the browse and dashboard views are rendered and updated, now by sending JSON data to the browser and creating the tables dynamically.  The client side remembers its current state when refreshing, and you can choose to persist this state with a save button when you have a layout you like.  The layout consists of the collapsed state of templated projects and groups.
On the backend some optimisations have been made for the data lookup: both a new database index and caching for build results if they are looked up twice because the project appears in two separate groups.  The latter is not likely to be a huge performance boost but is necessary for consistency anyway.

This change also contains fixes for IE8 compatibility (CIB-1947), including a (thankfully quite simple) upgrade to Ext 2.2.1.